requirements.txt
black==22.10.0
click==8.1.3
colorama==0.4.6
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.3
psycopg2-binary==2.9.5
python-decouple==3.6
tomli==2.0.1
typing_extensions==4.4.0
